Try to fix crashes on boot

+++ Virtual/Machine/ISO-maker/Files/For-Target/start.sh
@@ -32,11 +32,12 @@ mkdir -p root/.config/rclone
 mkdir -p root/etc/systemd/system/getty@tty1.service.d  # /override.conf Allows autologin to console as econ-ark
 mkdir -p root/usr/share/lightdm/lightdm.conf.d         # Configure display manager 
 
-# These items are created in econ-ark.seed; put them in /var/local so all system mods are findable there 
-ln -s /etc/rc.local                                          /var/local/root/etc
-ln -s /etc/default/grub                                      /var/local/root/etc/default
-ln -s /etc/systemd/system/getty@tty1.service.d/override.conf /var/local/root/etc/systemd/system/getty@tty1.service.d
-ln -s /usr/share/lightdm/lightdm.conf.d                      /var/local/root/usr/share/lightdm
+# These items are created in econ-ark.seed; put them in /var/local so all system mods are findable there
+# The ! -e are there in case the script is being rerun after a first install
+[[ ! -e /var/local/rc.local                                     ]] && ln -s /etc/rc.local                                          /var/local/root/etc
+[[ ! -e /var/local/root/etc/default    			        ]] && ln -s /etc/default/grub                                      /var/local/root/etc/default    
+[[ ! -e /var/local/root/etc/systemd/system/getty@tty1.service.d ]] && ln -s /etc/systemd/system/getty@tty1.service.d/override.conf /var/local/root/etc/systemd/system/getty@tty1.service.d
+[[ ! -e /var/local/root/usr/share/lightdm                       ]] && ln -s /usr/share/lightdm/lightdm.conf.d                      /var/local/root/usr/share/lightdm                      
 
 msg="$(cat ./About_This_Install/commit-msg.txt)"
 short_hash="$(cat ./About_This_Install/short.git-hash)"
